frontend: visitParty contributions implementation started, backend adjusted

+++ backend/services/creation/food_requirement/domains/model_food_requirement.go
@@ -11,7 +11,7 @@ type FoodRequirement struct {
 	PartyID        uint          `json:"party_id"`
 	Party          domains.Party `json:"-"`
 	Type           string        `json:"type" validate:"required"`
-	TargetQuantity int           `json:"target_quantity" validate:"required,gt=0"`
+	TargetQuantity float32       `json:"target_quantity" validate:"required,gt=0"`
 	QuantityMark   string        `json:"quantity_mark" validate:"required"`
 	Description    string        `json:"description"`
 }
